compute-client: move command history into its own module
This commit moves the `ComputeCommandHistory` type into its own module.
The `protcol::command` module now only contains the compute commands and
their documentation.
